# Conflicts:
#	Simulator/src/allow/simulator/entity/Person.java
#	Simulator/src/allow/simulator/flow/activity/person/PlanJourney.java
#	Simulator/src/allow/simulator/knowledge/EvoKnowledge.java
#	Simulator/src/allow/simulator/netlogo/agent/NetLogoWrapper.java
#	Simulator/src/allow/simulator/netlogo/commands/ExchangeKnowledge.java
#	Simulator/src/allow/simulator/netlogo/commands/SetupSimulator.java
#	Simulator/src/allow/simulator/netlogo/commands/Tick.java
#	Simulator/src/allow/simulator/statistics/Statistics.java
#	Simulator/src/allow/simulator/world/overlay/RasterOverlay.java

+++ Simulator/src/allow/simulator/entity/BusAgency.java
@@ -22,7 +22,7 @@ public final class BusAgency extends TransportationAgency {
 	// "Live" information about current trips and vehicles executing trips
 	protected final Map<String, TransportationEntity> currentlyUsedVehicles;
 	
-	public BusAgency(long id, Context context, String agencyId) {
+	public BusAgency(int id, Context context, String agencyId) {
 		super(id, context, agencyId);
 		routes = new HashMap<String, Route>();
 		nextTrips = new LinkedList<PublicTransportationTrip>();
